[#76m2x3] Displaying requests with status need refill.
