+++ Bionessori.Core/BaseRequest.cs
@@ -19,5 +19,35 @@ namespace Bionessori.Core {
         /// </summary>
         /// <returns>Список с заявками.</returns>
         public abstract Task<IEnumerable> GetRequests();
+
+        /// <summary>
+        /// Метод получает список заявок со статусом "Новая".
+        /// </summary>
+        /// <returns></returns>
+        public abstract Task<IEnumerable> GetDynamicDataNewRequests();
+
+        /// <summary>
+        /// Метод получает список заявок со статусом "В работе".
+        /// </summary>
+        /// <returns></returns>
+        public abstract Task<IEnumerable> GetDynamicDataWorkRequests();
+
+        /// <summary>
+        /// Метод получает список заявок со статусом "Требует подтверждения удаления".
+        /// </summary>
+        /// <returns></returns>
+        public abstract Task<IEnumerable> GetDynamicDataAcceptDeleteRequests();
+
+        /// <summary>
+        /// Метод получает материалы требующие пополнения.
+        /// </summary>
+        /// <returns></returns>
+        public abstract Task<IEnumerable> GetDynamicDataRefillMaterials();
+
+        /// <summary>
+        /// Метод получает материалы требующие сопоставления.
+        /// </summary>
+        /// <returns></returns>
+        public abstract Task<IEnumerable> GetDynamicDataMappingMaterials();
     }
 }
