--- Bionessori/Views/Route/RouteDynamicStructureData.cshtml
@@ -1,143 +1,0 @@
-﻿@{
-	Layout = "_Layout";
-}
-
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-	<title>@ViewData["Title"]</title>
-	<link href="~/css/Header.css" rel="stylesheet" />
-	<link href="~/css/Main.css" rel="stylesheet" />
-	<link href="~/css/Footer.css" rel="stylesheet" />
-</head>
-<body>
-	<script src="https://cdn.jsdelivr.net/npm/vue"></script>
-	<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-	<script src="https://use.fontawesome.com/1c903a9a3d.js"></script>
-
-	<div id="main_mrp">
-		<div>
-			<div class="col-lg-12">
-				<div class="row block-main-info">
-					<div class="btn-block-main mb-3 col-lg-12">
-						<!-- Левая секция кнопок -->
-						<div class="btn-left-block">
-							<div class="btns">
-								<nav aria-label="breadcrumb">
-									<ol class="breadcrumb">
-										<li class="breadcrumb-item"><a href="@Url.Action("Index", "Route")">Главная</a></li>
-
-										<li class="breadcrumb-item">
-											<a href="@Url.Action("RouteMRP", "Route")">Планирование потребностей в материалах</a>
-										</li>
-
-										<li class="breadcrumb-item active" aria-current="page">Конкретный список</li>
-									</ol>
-								</nav>
-							</div>
-						</div>
-
-						<!-- Правая секция кнопок -->
-						<div class="btn-right-block">
-							@*<button type="button" class="btn-item btn btn-default btn-nomenclature" v-on:click="">Отчеты</button>*@
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
-
-		<!-- Секция с таблицей с конкретным списком. -->
-		<div class="row">
-			<div class="col-lg-12">
-				<h3>Получение списка</h3>
-			</div>
-		</div>
-
-		@*<div class="tbl">
-			<table class="table table-bordered">
-				<thead>
-					<tr>
-						<th class="text-center" scope="col">
-							<button class="btn btn-link" value="" v-on:click="">
-								Пациент
-							</button>
-						</th>
-
-						<th class="text-center" scope="col">
-							<button class="btn btn-link" value="" v-on:click="">
-								Место направления
-							</button>
-						</th>
-
-						<th class="text-center" scope="col">
-							<button class="btn btn-link" value="" v-on:click="">
-								Тип направления
-							</button>
-						</th>
-
-						<th class="text-center" scope="col">
-							<button class="btn btn-link" value="" v-on:click="">
-								Статус направления
-							</button>
-						</th>
-
-						<th class="text-center" scope="col">
-							<button class="btn btn-link" value="" v-on:click="">
-								Специалист
-							</button>
-						</th>
-
-						<th class="text-center" scope="col"></th>
-						<th class="text-center" scope="col"></th>
-					</tr>
-				</thead>
-
-				<tbody>
-					<tr id="tbl-item tbl-row-click" class="tbl-item" v-for="el in aDirections">
-						<td class="text-center req-id">
-
-						</td>
-
-						<td class="text-center">
-							<span></span>
-						</td>
-
-						<td class="text-center">
-							<span></span>
-						</td>
-
-						<td class="text-center">
-							<span></span>
-						</td>
-
-						<td class="text-center">
-							<span></span>
-						</td>
-
-						<td class="text-center">
-							<span></span>
-						</td>
-
-						<td class="text-center state">
-							<button type="button" id="id-icon-edit" class="btn btn-default btn-sm" value="edit_direction" v-on:click="onRouteEditDirection">
-								<i class="fa fa-pencil" aria-hidden="true"></i>
-							</button>
-						</td>
-
-						<td class="text-center state">
-							<button type="button" id="id-icon-delete" class="btn btn-default btn-sm" data-toggle="modal" data-target="#delete-reception-modal" v-on:click="onDeleteDirection">
-								<i class="fa fa-trash-o" aria-hidden="true"></i>
-							</button>
-						</td>
-					</tr>
-				</tbody>
-			</table>
-		</div>*@
-	</div>
-
-	<script src="~/js/Route.js"></script>
-	<script src="~/js/MRP.js"></script>
-	<script src="~/lib/jquery/dist/jquery.min.js"></script>
-	<script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-</body>
-</html>
