+++ Bionessori/Views/Shared/_PartialRefillMaterials.cshtml
@@ -46,7 +46,7 @@
 			</div>
 		</div>
 
-		<!-- Секция материалами, которые нужно пополнить. -->
+		<!-- Секция с материалами, которые нужно пополнить. -->
 		<div class="row">
 			<div class="col-lg-12">
 				<h3>Материалы для пополнения</h3>
@@ -84,7 +84,7 @@
 				</thead>
 
 				<tbody>
-					<tr id="tbl-item tbl-row-click" class="tbl-item" v-for="el in aInWorkRequests">
+					<tr id="tbl-item tbl-row-click" class="tbl-item" v-for="el in aRefillMaterials">
 						<td class="text-center req-id">
 							{{el.id}}
 						</td>
